requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.0.8
django-cache-url==3.4.2
django-environ==0.9.0
marshmallow==3.19.0
environs==9.5.0
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dotenv==0.21.0
sqlparse==0.4.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
pandas==1.5.3
py3Dmol==2.0.1.post1
XlsxWriter==3.1.0
xlrd==2.0.1
python-decouple==3.8
gunicorn==20.1.0
django-debug-toolbar==4.0.0
pycodestyle==2.10.0
djlint==1.25.0
django-select2==8.1.2
djangorestframework==3.14.0
drf-yasg
drf-yasg[validation]
newrelic==9.1.0
django-filter
django-jsonapi